tests/requirements.txt
nextflow
pyyaml